adding id else statement for /start issue

+++ bot.py
@@ -1,11 +1,7 @@
-from telegram import *
 from telegram.ext import Updater, MessageHandler, Filters
 import os
-
 import utils
 from loguru import logger
-import requests
-
 
 class Bot:
 
@@ -37,6 +33,7 @@ class Bot:
 
 
 class QuoteBot(Bot):
+
     def _message_handler(self, update, context):
         to_quote = True
 
@@ -45,17 +42,18 @@ class QuoteBot(Bot):
 
         self.send_text(update, f'Your original message: {update.message.text}', quote=to_quote)
 
-
 class YoutubeBot(Bot):
 
     def _message_handler(self, update, context):
-        self.send_text(update, f'Wait please your video is dowloading : {update.message.text}')
-        downloaded_videos = utils.search_download_youtube_video(update.message.text, num_results=1)
-        for index, video in enumerate(downloaded_videos, start=1):
-            self.send_text(update, f'Video {index}/{len(downloaded_videos)}')
-            context.bot.send_video(update.message.chat_id, open(video , 'rb'), True)
-            os.remove(f'./{video}')
-
+        if update.message.text == "/start":
+            self.send_text(update, "I am ready, please enter the video name!")
+        else:
+            self.send_text(update, f'Wait please your video is dowloading : {update.message.text}')
+            downloaded_videos = utils.search_download_youtube_video(update.message.text, num_results=1)
+            for index, video in enumerate(downloaded_videos, start=1):
+                self.send_text(update, f'Video {index}/{len(downloaded_videos)}')
+                context.bot.send_video(update.message.chat_id, open(video, 'rb'), True)
+                os.remove(f'./{video}')
 
 if __name__ == '__main__':
     with open('secret/.telegramToken') as f:
